Creates sd-viewer metapackage, replaces sd-svs-disp
Adds new metapackage with a clean history, to aid in template
consolidation. Attempts to override the old package, causing it
automatically be removed, via Replaces/Conflicts [0].

The dependencies are the same, although reordered, with one exception:
"evince" has been pulled in so we can remove it from the salt config.

[0] https://www.debian.org/doc/debian-policy/ch-relationships.html#s-replaces